refactor: move public types to types.ts, drop redundant overload
