+++ src/index.ts
@@ -5,9 +5,8 @@ import { OutputFnType } from './types'
 import createPcg from './createPcg'
 
 export { stepState, nextState, prevState, randomInt, randomList } from './createPcg'
-export type { CreatePcg, CreatePcgOptions, LongLike, RandomFn } from './createPcg'
 export { OutputFnType, StreamScheme } from './types'
-export type { OutputFn, PCGConfig, PCGState, SchemeFn } from './types'
+export type { CreatePcg, CreatePcgOptions, LongLike, OutputFn, PCGConfig, PCGState, RandomFn, SchemeFn } from './types'
 
 export const createPcg32 = createPcg({
   numOutputBits: 32,
